+++ CHANGES.md
@@ -7,11 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- (BREAKING) Remove `Sidekiq::Antidote::Config#key_prefix`
+
+
 ## [1.0.0.alpha.1] - 2023-12-01
 
 ### Added
 
-* Initial release.
+- Initial release.
 
 
+[unreleased]: https://github.com/ixti/sidekiq-pauzer/compare/v1.0.0.alpha.1...main
 [1.0.0.alpha.1]: https://github.com/ixti/sidekiq-antidote/tree/v1.0.0.alpha.1